feat: let a goal be changed, and show what changing it costs
The affordability line shipped an hour ago tells a user their goals cost more
than they earn, that a target or date has to move, and that pushing a date out
is usually cheaper than cutting a target. Goals could be created and read and
nothing else. Advice with the door locked.

PATCH and DELETE now exist, and everything a goal's inputs decide is recomputed
rather than patched — change the date and the monthly SIP, the equity split and
the fund plan all follow, and the explanation is regenerated instead of left
describing the plan the goal used to have. Changing the goal type moves its
inflation rate with it unless one was pinned explicitly.

The edit form previews the new monthly figure live, because "pushing the date
out is cheaper" is a claim the user should be able to check rather than take on
trust. Three things had to be true for that preview to be honest, and none of
them were:

- The preview asked for a whole number of years while the save sent 16.8, so it
  quoted ₹13,218 and stored ₹14,501. It now sends exactly what the save sends.
- The SIP calculator truncated years to an integer. A goal 16.8 years away was
  priced over 16, quietly asking for a larger monthly figure than the goal
  needs. The arithmetic is continuous in it; the int() cast was the only thing
  forcing the rounding.
- When the preview failed it fell back to the current figure, which reads as
  "nothing changes" — the one thing a broken preview must not say. It now says
  it could not compute.

Deleting drops the goal's queries rather than invalidating them. Invalidating a
deleted goal refetches it, and the 404 rendered "couldn't load this goal" over a
page already on its way out.

+++ backend/app/schemas/advisor.py
@@ -5,7 +5,9 @@ from pydantic import BaseModel, Field
 
 class SipRequest(BaseModel):
     target_amount: float
-    years: int
+    # Fractional on purpose: a goal 16.8 years away is priced over 16.8 years,
+    # and the edit form's live preview sends exactly what the save will send.
+    years: float
     annual_return_rate: float = 0.12
     current_savings: float = 0.0
     inflation_rate: float = 0.06
